# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Server API",
        "title": "Golang",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:1320",
    "basePath": "/",
    "paths": {
        "/auth/register": {
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "summary": "node register",
                "parameters": [
                    {
                        "description": "payload",
                        "name": "payload",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.NodeCreateRequest"
                        }
                    }
                ],
                "responses": {}
            }
        }
    },
    "definitions": {
        "dto.NodeCreateRequest": {
            "type": "object",
            "properties": {
                "confirm_password": {
                    "type": "string",
                    "example": "123456"
                },
                "email": {
                    "type": "string",
                    "example": "mehrdad@gmail.com"
                },
                "password": {
                    "type": "string",
                    "example": "123456"
                },
                "phone_number": {
                    "type": "string",
                    "example": "09120246217"
                },
                "referral": {
                    "type": "string",
                    "example": "L100@2154"
                },
                "username": {
                    "type": "string",
                    "example": "mehrdad"
                },
                "wallet_id": {
                    "type": "string",
                    "example": "TZJjSqfXwTP5YgjdhLdRV1zqpDwDsgdZZE"
                }
            }
        }
    }
}